added directory layout to huff modules
